+++ guillotina_cms/interfaces/registry.py
@@ -5,17 +5,35 @@ import json
 
 MENU_LAYOUT = json.dumps({"type": "object", "properties": {}})
 
+LAYOUT_TYPE_COMPONENTS = json.dumps(
+    {
+        "type": "object",
+        "properties": {},
+        "additionalProperties": {
+            "type": "array",
+            "items": {
+                "type": "object",
+                "properties": {
+                    "id": {"type" : "string"}
+                }
+            },
+        },
+    }
+)
+
+
 class IImagingSettings(Interface):
     allowed_sizes = schema.Dict(
         missing_value={
-            'high': '1400:1400',
-            'large': '768:768',
-            'preview': '400:400',
-            'mini': '200:200',
-            'thumb': '128:128',
-            'tile': '64:64',
-            'icon': '32:32'
-        })
+            "high": "1400:1400",
+            "large": "768:768",
+            "preview": "400:400",
+            "mini": "200:200",
+            "thumb": "128:128",
+            "tile": "64:64",
+            "icon": "32:32",
+        }
+    )
 
     quality = schema.Int(default=88)
 
@@ -23,14 +41,22 @@ class IImagingSettings(Interface):
 class IMenu(Interface):
 
     definition = schema.JSONField(
-        title="Menu definition",
-        required=False,
-        schema=MENU_LAYOUT,
-        defaultFactory=list
+        title="Menu definition", required=False, schema=MENU_LAYOUT, defaultFactory=list
     )
 
-    logo = schema.Text(
-        title="Logo",
+    logo = schema.Text(title="Logo", required=False, defaultFactory=get_default_logo)
+
+
+class ICustomTheme(Interface):
+
+    css = schema.Text(title="CSS Text", required=False, default="")
+
+
+class ILayoutComponents(Interface):
+
+    components = schema.JSONField(
+        title="Layout enabled components by type",
         required=False,
-        defaultFactory=get_default_logo
+        schema=LAYOUT_TYPE_COMPONENTS,
+        defaultFactory=dict,
     )